requirements.txt
# Django & DRF
Django==5.0.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1

# API Documentation
drf-spectacular==0.27.1

# Environment & Configuration
python-dotenv==1.0.1

# CORS
django-cors-headers==4.3.1

# Filtering & Search
django-filter==23.5

# Image Processing
Pillow==10.2.0

# Database
psycopg2-binary==2.9.9

# Security & Rate Limiting
django-ratelimit==4.1.0

# Testing
pytest==7.4.4
pytest-django==4.7.0
pytest-cov==4.1.0

# Development Tools
django-extensions==3.2.3
